// types/ViewType.ts
export type IViewType =
  | 'full-width-banner'
  | 'plp-box'
  | 'banner-with-side-fields-plus-related'
  | 'default'
  | 'text-module'
  | 'banner-with-side-fields'
  | 'blurred-slider'
  | 'landscape-banner'
  | 'narrow-banner-w-all-fields-below'
  | 'box-plus-one'
  | 'box-with-margin'
  | 'boards-with-fields'
  | 'boards-with-fields-on-below-mobile-no-carousel'
  | 'boards-with-fields-on-below-mobile-carousel'
  | 'boards-with-fields-below'
  | 'grid-of-boards-two-columns'
  | 'grid-of-boards-three-columns'
  | 'combo-mini-slider-plus-box-all-fields'
  | 'products'
  | 'dcw-products'
  | 'grid-of-products'
  | 'mini-slider'
  | 'open-article_without_title'
  | 'categories-with-icon'
  | 'group-of-menu-items'
  | 'big-menu-banner'
  | 'banners-with-icons'
  | 'list-of-products'
  | 'list-of-brand-icons'
  | 'list-of-brands'
  | 'book-an-appointment-plus-search-box'
  | 'plp-one-tile'
  | 'plp-two-tiles'
  | 'plp-three-tiles'
  | 'squat-banner'
  | 'top-page-banner'
  | 'square-boards-with-split'
  | 'square-boards-without-split'
  | 'faqs'
  | 'faq-search'
  | 'you-may-also-like'
  | 'size-guide'
  | 'query-list'
  | 'wall-of-brands'
  | 'Algolia Recommendations'
  | 'Webform'
  | 'Newsletter'
  | 'open-article'
  | null